cmd/go-cache-plugin: use TCP loopback instead of a Unix socket
A Unix-domain socket does not mesh well when the server may run as a different
user from the Go toolchain. Specifically, connecting to a Unix socket requires
the caller to have write permission to the socket, and it is tricky (and OS
dependent) to chmod a socket once created.

So, instead use a loopback TCP socket. Update the environment, help text, and
flags to use "--plugin" and "GOCACHE_PLUGIN" instead of "socket". Update the
connect command to accept a port instead of a path.